Allow users to be created without a created-by clientId
* Fixes a typo in the dss-database main changelog

+++ pepper-apis/dss-core/src/test/java/org/broadinstitute/ddp/db/dao/UserGovernanceDaoTest.java
@@ -36,7 +36,7 @@ public class UserGovernanceDaoTest extends TxnAwareBaseTest {
 
             User governed = handle.attach(UserDao.class).findUserById(gov.getGovernedUserId()).get();
             assertNull(governed.getAuth0UserId());
-            assertEquals(testData.getClientId(), governed.getCreatedByClientId());
+            assertEquals(testData.getClientId(), governed.getCreatedByClientId().longValue());
 
             handle.rollback();
         });
@@ -55,7 +55,7 @@ public class UserGovernanceDaoTest extends TxnAwareBaseTest {
 
             User governed = handle.attach(UserDao.class).findUserById(gov.getGovernedUserId()).get();
             assertNull(governed.getAuth0UserId());
-            assertEquals(testData.getClientId(), governed.getCreatedByClientId());
+            assertEquals(testData.getClientId(), governed.getCreatedByClientId().longValue());
 
             handle.rollback();
         });
